+++ src/gravity/types.ts
@@ -1,0 +1,5 @@
+import type { Vector3 } from "@babylonjs/core/Maths/math.vector";
+
+export interface GravityInterface {
+  getGravity(getCenterOfMass: ()=>Vector3): Vector3;
+}